refactor: apply barrel export pattern and clean up import paths
- Add barrel exports (index.ts) to src/cli/ and src/internal/utils/ directories
- Enhance existing barrel exports (errors, types, pull) with re-exports
- Remove all .js extensions from TypeScript files for cleaner code
- Unify import paths to directory-level (e.g., from "./logger.js" → from "./utils")

fix #100
