+++ src/cli/index.ts
@@ -1,0 +1,2 @@
+export type { PullCLIOptions } from "./types";
+export { adaptCLIOptions } from "./adapter";
