+++ src/index.ts
@@ -1,2 +1,2 @@
-export { baedal } from "./pkg/pull/index.js";
-export type { BaedalOptions, PullResult } from "./internal/types/index.js";
+export type { BaedalOptions, PullResult } from "./internal/types";
+export { baedal } from "./pkg/pull";
