+++ src/internal/errors/network.ts
@@ -1,4 +1,4 @@
-import { BaseError } from "./base.js";
+import { BaseError } from "./base";
 
 export class NetworkError extends BaseError {
   constructor(
