+++ src/pkg/push/files.spec.ts
@@ -1,8 +1,8 @@
 import { mkdtemp, mkdir, rm, writeFile } from "node:fs/promises";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
-import { FileSystemError } from "../../internal/errors/index.js";
-import { collectFiles, getRelativePath, normalizePath } from "./files.js";
+import { FileSystemError } from "../../internal/errors";
+import { collectFiles, getRelativePath, normalizePath } from "./files";
 
 describe("Push Files - Integration Test", () => {
   const MAX_FILE_SIZE_MB = 11;
