+++ src/pkg/push/index.ts
@@ -1,6 +1,7 @@
-export { loadPushConfig } from "./config.js";
-export { executePush } from "./executor.js";
-export { collectFiles } from "./files.js";
-export { createGitHubClient } from "./github.js";
-export { initPushConfig, printInitSuccess } from "./init.js";
-export type * from "./types.js";
+export { loadPushConfig } from "./config";
+export { executePush } from "./executor";
+export { collectFiles } from "./files";
+export { createGitHubClient } from "./github";
+export { initPushConfig, printInitSuccess } from "./init";
+export { PushConfigSchema } from "./types";
+export type * from "./types";
